working on cloudinary image upload still

+++ frontend/src/pages/admin/AdminEditProductPage.js
@@ -32,7 +32,12 @@ const AdminEditProductPage = () => {
 
   const imageDeleteHandler = async (imagePath, productId) => {
       let encoded = encodeURIComponent(imagePath)
+      if (process.env.NODE_ENV === "production") {
+          // to do: change to !==
       await axios.delete(`/api/products/admin/image/${encoded}/${productId}`);
+      } else {
+        await axios.delete(`/api/products/admin/image/${encoded}/${productId}?cloudinary=true`);  
+      }
   }
 
   return <EditProductPageComponent categories={categories} fetchProduct={fetchProduct} updateProductApiRequest={updateProductApiRequest} reduxDispatch={reduxDispatch} saveAttributeToCatDoc={saveAttributeToCatDoc} imageDeleteHandler={imageDeleteHandler} uploadHandler={uploadHandler} />;
